+++ OrthancServer/Sources/ServerContext.cpp
@@ -361,11 +361,12 @@ namespace Orthanc
   {
     metricsRegistry_->SetFloatValue("orthanc_dicom_cache_size_mb",
                                     static_cast<float>(dicomCache_.GetCurrentSize()) / static_cast<float>(1024 * 1024));
-    metricsRegistry_->SetIntegerValue("orthanc_dicom_cache_count", dicomCache_.GetNumberOfItems());
-
+    metricsRegistry_->SetIntegerValue("orthanc_dicom_cache_count", 
+                                    static_cast<int64_t>(dicomCache_.GetNumberOfItems()));
     metricsRegistry_->SetFloatValue("orthanc_storage_cache_size_mb",
                                     static_cast<float>(storageCache_.GetCurrentSize()) / static_cast<float>(1024 * 1024));
-    metricsRegistry_->SetIntegerValue("orthanc_storage_cache_count", storageCache_.GetNumberOfItems());
+    metricsRegistry_->SetIntegerValue("orthanc_storage_cache_count", 
+                                    static_cast<int64_t>(storageCache_.GetNumberOfItems()));
   }
